Add support for Django 1.10 new style of middleware.
Update the tox env for Django 1.10 tests for python 2.7, 3.4, and 3.5

Try adding the `__call__` and `__init__` methods manually.

Use the deprecation `MiddlewareMixin` rather than calling `__call__` and
`__init__`.

Add a guard for the importing the `deprecation.MiddlewareMixin`

Add django 1.10 to the `.travis.yml` file.

Add a new line above the `try` for style reasons.