Remove upper bound on Django version
Django's backwards compatibility and deprecation policies mean that any future compatibility issues will be known well in advance of any breakage. There is no advantage to pinning the upper bound on the supported Django version. It means a django-csp release is required to support each version of Django (and we're currently waiting on a django-csp release to upgrade to Django 1.11).