Allow reading nonce if it was included in header
If the nonce was generated before the CSP headers were set, then allow
reading it with request.csp_nonce.

If the CSP headers were set with no nonce, then continue raising
CSPNonceError when reading it as a string. If read as a boolean, then
return False.

This will allow other middleware like django-debug-toolbar to alter the
response after the CSP middleware runs.